Log-Rank Test for Internal Calibration and External Calibration Results

Log-Rank Test for Internal Calibration and External Calibration Results

hdnom.logrank(object)

Arguments

object

An object returned by hdnom.calibrate or hdnom.external.calibrate.

Examples

library("survival")

# Load imputed SMART data
data("smart")

# Use the first 1000 samples as training data
# (the data used for internal validation)
x = as.matrix(smart[, -c(1, 2)])[1:1000, ]
time = smart$TEVENT[1:1000]
event = smart$EVENT[1:1000]

# Take the next 1000 samples as external calibration data
# In practice, usually use data collected in other studies
x_new = as.matrix(smart[, -c(1, 2)])[1001:2000, ]
time_new = smart$TEVENT[1001:2000]
event_new = smart$EVENT[1001:2000]

# Fit Cox model by lasso penalization
lassofit = hdcox.lasso(x, Surv(time, event), nfolds = 5, rule = "lambda.1se", seed = 11)

### Internal calibration
cal.int = hdnom.calibrate(x, time, event, model.type = "lasso",
                          alpha = 1, lambda = lassofit$'lasso_best_lambda',
                          method = "cv", nfolds = 5,
                          pred.at = 365 * 9, ngroup = 3)

#> Start fold 1 
#> Start fold 2 
#> Start fold 3 
#> Start fold 4 
#> Start fold 5 

hdnom.logrank(cal.int)

#> Call:
#> survdiff(formula = formula("Surv(time, event) ~ grp"))
#> 
#> n=999, 1 observation deleted due to missingness.
#> 
#>         N Observed Expected (O-E)^2/E (O-E)^2/V
#> grp=1 333      121     67.0     43.45     62.07
#> grp=2 333       63     77.3      2.63      4.01
#> grp=3 333       41     80.7     19.54     30.53
#> 
#>  Chisq= 65.8  on 2 degrees of freedom, p= 5.11e-15 

### External calibration
cal.ext =
  hdnom.external.calibrate(lassofit, x, time, event,
                           x_new, time_new, event_new,
                           pred.at = 365 * 5, ngroup = 3)

hdnom.logrank(cal.ext)

#> Call:
#> survdiff(formula = formula("Surv(time, event) ~ grp"))
#> 
#> n=999, 1 observation deleted due to missingness.
#> 
#>         N Observed Expected (O-E)^2/E (O-E)^2/V
#> grp=1 339       80     46.7     23.69     34.90
#> grp=2 327       40     48.0      1.35      2.01
#> grp=3 333       26     51.2     12.42     19.15
#> 
#>  Chisq= 37.5  on 2 degrees of freedom, p= 7.12e-09